refactor(model-providers): extract modelDefaults router logic into service
Move the fat tRPC handlers (getDefaultModelsForProject, getInheritedValuesForScopes,
getResolvedDefault) into a new modelDefaults.read.ts module exporting
getDefaultModelsSnapshot, getInheritedValuesForScopes, getResolvedDefaultForFeature.
Add assertCanWriteScope + getScopeAttachmentsForConfig to modelDefaults.service.ts
so the auth gate and config-scope lookups are reusable. The tRPC router now
only does zod parsing + tRPC permission middleware and delegates everything
else to the service layer, so the same surface can be called from a Hono
endpoint (and CLI/API clients) without duplicating business logic.

Also drop an unused Prisma import flagged by github-code-quality.
